requirements.txt
python-telegram-bot==20.7
fastapi==0.109.0
uvicorn==0.25.0
google-api-python-client==2.114.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
openai==1.10.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
aiofiles==23.2.1
httpx~=0.25.2